refactor: isolate Radio controls and tuner presentation
