+++ package.json
@@ -129,6 +129,7 @@
     "./ui/maps": "./src/ui/mapSource.js",
     "./ui/layers": "./src/ui/layers.js",
     "./ui/layers/feedback": "./src/loadingFeedback.js",
-    "./ui/location": "./src/ui/location.js"
+    "./ui/location": "./src/ui/location.js",
+    "./ui/radio": "./src/ui/radio.js"
   }
 }
