+++ scripts/format-scope.json
@@ -166,5 +166,11 @@
   "src/locationStatus.js",
   "src/ui/locationControls.test.mjs",
   "src/ui/locationSearch.test.mjs",
-  "scripts/qa-location-controls.mjs"
+  "scripts/qa-location-controls.mjs",
+  "src/ui/radio.js",
+  "src/ui/radioControls.js",
+  "src/ui/radioBindings.js",
+  "src/ui/radioPresentation.js",
+  "src/ui/radioTunerModel.js",
+  "src/ui/radioControls.test.mjs"
 ]
